requirements.txt
colorama==0.4.6
einops==0.8.1
icecream==2.1.8
loguru==0.7.3
numpy==2.3.5
openai==2.9.0
Pillow==12.0.0
pydantic==2.12.5
python-dotenv==1.2.1
PyYAML==6.0.2
PyYAML==6.0.3
qwen_vl_utils==0.0.14
Requests==2.32.5
sentence_transformers==5.1.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.0
